Implement closers for Jaeger container.
