+++ pkg/tests/end_to_end_tests/ingest_trace_test.go
@@ -6,8 +6,6 @@ package end_to_end_tests
 
 import (
 	"context"
-	"fmt"
-	"math/rand"
 	"testing"
 	"time"
 
@@ -17,7 +15,6 @@ import (
 	"github.com/timescale/promscale/pkg/jaeger/query"
 	ingstr "github.com/timescale/promscale/pkg/pgmodel/ingestor"
 	"github.com/timescale/promscale/pkg/pgxconn"
-	"go.opentelemetry.io/collector/model/pdata"
 )
 
 func TestIngestTraces(t *testing.T) {
@@ -42,151 +39,6 @@ func TestIngestTracesMultiTraces(t *testing.T) {
 		}
 		require.NoError(t, err)
 	})
-}
-
-func generateTestTrace() pdata.Traces {
-	spanCount := 4
-	td := pdata.NewTraces()
-	rs := td.ResourceSpans().AppendEmpty()
-	initResourceAttributes(rs.Resource().Attributes(), 0)
-	libSpans := rs.InstrumentationLibrarySpans().AppendEmpty()
-	initInstLib(libSpans, 0)
-	for i := 0; i < spanCount; i++ {
-		fillSpanOne(libSpans.Spans().AppendEmpty())
-	}
-	ids := make([]pdata.SpanID, 0)
-	for i := 0; i < spanCount; i++ {
-		sid := pdata.NewSpanID(generateRandSpanID())
-		fillSpanTwo(libSpans.Spans().AppendEmpty(), sid)
-		ids = append(ids, sid)
-	}
-	rs = td.ResourceSpans().AppendEmpty()
-	initResourceAttributes(rs.Resource().Attributes(), 1)
-	libSpans = rs.InstrumentationLibrarySpans().AppendEmpty()
-	initInstLib(libSpans, 1)
-	for _, parentID := range ids {
-		fillSpanThree(libSpans.Spans().AppendEmpty(), parentID)
-	}
-
-	return td
-}
-
-func generateTestTraceManyRS() []pdata.Traces {
-	traces := make([]pdata.Traces, 5)
-	for traceIndex := 0; traceIndex < len(traces); traceIndex++ {
-		spanCount := 5
-		td := pdata.NewTraces()
-		for resourceIndex := 0; resourceIndex < 5; resourceIndex++ {
-			rs := td.ResourceSpans().AppendEmpty()
-			initResourceAttributes(rs.Resource().Attributes(), resourceIndex)
-			for libIndex := 0; libIndex < 5; libIndex++ {
-				libSpans := rs.InstrumentationLibrarySpans().AppendEmpty()
-				initInstLib(libSpans, libIndex)
-				for i := 0; i < spanCount; i++ {
-					fillSpanOne(libSpans.Spans().AppendEmpty())
-				}
-				for i := 0; i < spanCount; i++ {
-					fillSpanTwo(libSpans.Spans().AppendEmpty(), pdata.NewSpanID(generateRandSpanID()))
-				}
-			}
-		}
-		traces[traceIndex] = td
-	}
-	return traces
-}
-
-func initInstLib(dest pdata.InstrumentationLibrarySpans, index int) {
-	// Only add schema URL on half of instrumentation libs.
-	if index%2 == 0 {
-		dest.SetSchemaUrl(fmt.Sprintf("url-%d", index%2))
-	}
-	dest.InstrumentationLibrary().SetName(fmt.Sprintf("inst-lib-name-%d", index))
-	dest.InstrumentationLibrary().SetVersion(fmt.Sprintf("1.%d.0", index%2))
-}
-
-func initResourceAttributes(dest pdata.AttributeMap, index int) {
-	tmpl := pdata.NewAttributeMapFromMap(map[string]pdata.AttributeValue{
-		"resource-attr": pdata.NewAttributeValueString(fmt.Sprintf("resource-attr-val-%d", index%2)),
-		"service.name":  pdata.NewAttributeValueString(fmt.Sprintf("service-name-%d", index)),
-	})
-	dest.Clear()
-	tmpl.CopyTo(dest)
-}
-
-func initSpanEventAttributes(dest pdata.AttributeMap) {
-	dest.Clear()
-	spanEventAttributes.CopyTo(dest)
-}
-
-func initSpanAttributes(dest pdata.AttributeMap) {
-	dest.Clear()
-	spanAttributes.CopyTo(dest)
-}
-
-func initSpanLinkAttributes(dest pdata.AttributeMap) {
-	dest.Clear()
-	spanLinkAttributes.CopyTo(dest)
-}
-
-func generateRandSpanID() (result [8]byte) {
-	rand.Read(result[:])
-	return result
-}
-
-func fillSpanOne(span pdata.Span) {
-	span.SetTraceID(pdata.NewTraceID(traceID1))
-	span.SetSpanID(pdata.NewSpanID(generateRandSpanID()))
-	span.SetName("operationA")
-	span.SetStartTimestamp(testSpanStartTimestamp)
-	span.SetEndTimestamp(testSpanEndTimestamp)
-	span.SetDroppedAttributesCount(1)
-	span.SetTraceState("span-trace-state1")
-	initSpanAttributes(span.Attributes())
-	evs := span.Events()
-	ev0 := evs.AppendEmpty()
-	ev0.SetTimestamp(testSpanEventTimestamp)
-	ev0.SetName("event-with-attr")
-	initSpanEventAttributes(ev0.Attributes())
-	ev0.SetDroppedAttributesCount(2)
-	ev1 := evs.AppendEmpty()
-	ev1.SetTimestamp(testSpanEventTimestamp)
-	ev1.SetName("event")
-	ev1.SetDroppedAttributesCount(2)
-	span.SetDroppedEventsCount(1)
-	status := span.Status()
-	status.SetCode(pdata.StatusCodeError)
-	status.SetMessage("status-cancelled")
-}
-
-func fillSpanTwo(span pdata.Span, spanID pdata.SpanID) {
-	span.SetTraceID(pdata.NewTraceID(traceID2))
-	span.SetSpanID(spanID)
-	span.SetName("operationB")
-	span.SetStartTimestamp(testSpanStartTimestamp)
-	span.SetEndTimestamp(testSpanEndTimestamp)
-	span.SetTraceState("span-trace-state2")
-	initSpanAttributes(span.Attributes())
-	link0 := span.Links().AppendEmpty()
-	initSpanLinkAttributes(link0.Attributes())
-	link0.SetTraceID(pdata.NewTraceID([16]byte{'1'}))
-	link0.SetSpanID(pdata.NewSpanID(generateRandSpanID()))
-	link0.SetDroppedAttributesCount(4)
-	link0.SetTraceState("link-trace-state1")
-	link1 := span.Links().AppendEmpty()
-	link1.SetDroppedAttributesCount(4)
-	link1.SetTraceID(pdata.NewTraceID([16]byte{'1'}))
-	link1.SetSpanID(pdata.NewSpanID(generateRandSpanID()))
-	link1.SetTraceState("link-trace-state2")
-	span.SetDroppedLinksCount(3)
-}
-
-func fillSpanThree(span pdata.Span, parentSpanID pdata.SpanID) {
-	span.SetTraceID(pdata.NewTraceID(traceID2))
-	span.SetSpanID(pdata.NewSpanID(generateRandSpanID()))
-	span.SetParentSpanID(parentSpanID)
-	span.SetName("operationC")
-	span.SetStartTimestamp(testSpanStartTimestamp)
-	span.SetEndTimestamp(testSpanEndTimestamp)
 }
 
 func TestQueryTraces(t *testing.T) {
